+++ Example/Controllers/ViewController.swift
@@ -37,9 +37,6 @@ class ViewController: UIViewController {
     let locationManager = CLLocationManager()
     var permissionHandlerObj: PermissionCompletionHandler?
     var currentMiniAppTitle: String?
-    var messageHandlerObj: ChatContactHandler?
-    var messageIdHandlerObj: ChatContactHandler?
-    var messageMultipleHandlerObj: ChatContactsHandler?
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -58,7 +55,7 @@ class ViewController: UIViewController {
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         if segue.identifier == "DisplayMiniApp" {
             guard let miniAppDisplay = currentMiniAppView else {
-                displayAlert(title: NSLocalizedString("error_title", comment: ""), message: NSLocalizedString("error_miniapp_message", comment: ""), dismissController: true)
+                displayAlert(title: MASDKLocale.localize("miniapp.sdk.ios.error.title"), message: MASDKLocale.localize("miniapp.sdk.ios.error.message.miniapp"), dismissController: true)
                 return
             }
 
@@ -82,7 +79,7 @@ class ViewController: UIViewController {
                 case .success:
                     self.displayMiniApp(miniAppInfo: miniAppInfo)
                 case .failure(let error):
-                    self.displayAlert(title: NSLocalizedString("error_title", comment: ""), message: error.localizedDescription, dismissController: true)
+                    self.displayAlert(title: MASDKLocale.localize("miniapp.sdk.ios.error.title"), message: error.localizedDescription, dismissController: true)
                 }
             }
         } else {
@@ -100,7 +97,7 @@ class ViewController: UIViewController {
                     miniAppInfo: miniAppInfo,
                     customMetaData: manifestData.customMetaData ?? [:])
             case .failure:
-                self.displayAlert(title: NSLocalizedString("error_title", comment: ""), message: NSLocalizedString("error_single_message", comment: ""), dismissController: true)
+                self.displayAlert(title: MASDKLocale.localize("miniapp.sdk.ios.error.title"), message: MASDKLocale.localize("miniapp.sdk.ios.error.message.single"), dismissController: true)
             }
         }
     }