+++ Example/Utils/UIView+MiniApp.swift
@@ -27,3 +27,35 @@ extension UIButton {
        self.layer.borderColor = UIColor.white.cgColor
     }
 }
+
+@IBDesignable extension UIView {
+    @IBInspectable var borderColor: UIColor? {
+        get {
+            if let color = layer.borderColor {
+                return UIColor(cgColor: color)
+            } else {
+                return nil
+            }
+        }
+        set {
+            layer.borderColor = newValue!.cgColor
+        }
+    }
+    @IBInspectable var borderWidth: CGFloat {
+        get {
+            return layer.borderWidth
+        }
+        set {
+            layer.borderWidth = newValue
+        }
+    }
+    @IBInspectable var cornerRadius: CGFloat {
+        get {
+            return layer.cornerRadius
+        }
+        set {
+            layer.cornerRadius = newValue
+            clipsToBounds = newValue > 0
+        }
+    }
+}