+++ MiniApp/Classes/core/Display/ViewControllers/Custom Permissions/CustomPermissionsRequestViewController.swift
@@ -26,8 +26,9 @@ class CustomPermissionsRequestViewController: UIViewController {
         guard let permissionRequest = permissionsRequestList else {
             return
         }
-        customPermissionHandlerObj?(.success(permissionRequest))
-        self.dismiss(animated: true, completion: nil)
+        self.dismiss(animated: true, completion: {
+            self.customPermissionHandlerObj?(.success(permissionRequest))
+        })
     }
 
     @objc func permissionValueChanged(_ sender: UISwitch) {
@@ -43,14 +44,14 @@ class CustomPermissionsRequestViewController: UIViewController {
             $0.isPermissionGranted == MiniAppCustomPermissionGrantedStatus.denied
             } ?? false
         if allPermissionsDenied {
-            self.saveButton.title = "Save"
+            self.saveButton.title = MASDKLocale.localize(.save)
         } else {
-            self.saveButton.title = "Allow"
+            self.saveButton.title = MASDKLocale.localize(.allow)
         }
     }
 
     func addFooterInfo() {
-        self.footerLabel.text = " \(miniAppTitle) wants to access the above permissions. Choose your preference accordingly.\n\n  You can also manage these permissions later in the Miniapp settings"
+        self.footerLabel.text = MASDKLocale.localize(.firstLaunchFooter, miniAppTitle)
     }
 }