+++ MiniApp/Classes/core/JavascriptBridge/MiniAppJavaScriptEnumerations.swift
@@ -24,6 +24,7 @@ enum JavaScriptExecResult: String {
 enum MiniAppSupportedSchemes: String {
     case tel // used for phone calls
     case about // used to reveal internal state and built-in functions (e.g.: alert dialog)
+    case mailto
 }
 
 /// List of Device Permissions supported by the SDK that can be requested by a Mini app
@@ -42,6 +43,8 @@ public enum MiniAppCustomPermissionType: String, Codable, CaseIterable {
     case contactsList = "rakuten.miniapp.user.CONTACT_LIST"
     /// Custom permission for retrieving Access token details from the host app
     case accessToken = "rakuten.miniapp.user.ACCESS_TOKEN"
+    /// Custom permission to send a message via the Host app
+    case sendMessage = "rakuten.miniapp.user.action.SEND_MESSAGE"
     /// Custom permission for retrieving device location details from the host app
     case deviceLocation = "rakuten.miniapp.device.LOCATION"
 
@@ -57,6 +60,8 @@ public enum MiniAppCustomPermissionType: String, Codable, CaseIterable {
             return "Access Token"
         case .deviceLocation:
             return "Device Location"
+        case .sendMessage:
+            return "Send Message"
         }
     }
 }