+++ MiniApp/Classes/core/Protocols/MiniAppUserInfoDelegate.swift
@@ -11,6 +11,9 @@ public protocol MiniAppUserInfoDelegate: class {
     func getProfilePhoto(completionHandler: @escaping (Result<String?, MASDKError>) -> Void)
 
     /// Interface that is used to retrieve the Contact list
+    func getContacts(completionHandler: @escaping (Result<[MAContact]?, MASDKError>) -> Void)
+
+    @available(*, deprecated, renamed:"getContacts(completionHandler:)")
     func getContacts() -> [MAContact]?
 
     /// Interface that is used to retrieve the Token Info
@@ -26,15 +29,31 @@ public protocol MiniAppUserInfoDelegate: class {
 public extension MiniAppUserInfoDelegate {
 
     func getUserName(completionHandler: @escaping (Result<String?, MASDKError>) -> Void) {
-        completionHandler(.failure(.unknownError(domain: NSLocalizedString("host_app_error", comment: ""), code: 1, description: NSLocalizedString("failed_to_conform_to_protocol", comment: ""))))
+        completionHandler(.failure(.unknownError(domain: MASDKLocale.localize(.hostAppError), code: 1, description: MASDKLocale.localize(.failedToConformToProtocol))))
     }
 
     func getProfilePhoto(completionHandler: @escaping (Result<String?, MASDKError>) -> Void) {
-        completionHandler(.failure(.unknownError(domain: NSLocalizedString("host_app_error", comment: ""), code: 1, description: NSLocalizedString("failed_to_conform_to_protocol", comment: ""))))
+        completionHandler(.failure(.unknownError(domain: MASDKLocale.localize(.hostAppError), code: 1, description: MASDKLocale.localize(.failedToConformToProtocol))))
+    }
+
+    func getContacts(completionHandler: @escaping (Result<[MAContact]?, MASDKError>) -> Void) {
+        completionHandler(.failure(.unknownError(domain: MASDKLocale.localize(.hostAppError), code: 1, description: MASDKLocale.localize(.failedToConformToProtocol))))
     }
 
     func getContacts() -> [MAContact]? {
-        return nil
+        let semaphore = DispatchSemaphore(value: 0)
+        var contacts: [MAContact]?
+        getContacts { result in
+            switch result {
+            case .success(let listContacts):
+                contacts = listContacts
+            default:
+                contacts = nil
+            }
+            semaphore.signal()
+        }
+        semaphore.wait()
+        return contacts
     }
 
     func getAccessToken(miniAppId: String, scopes: MASDKAccessTokenScopes, completionHandler: @escaping (Result<MATokenInfo, MASDKCustomPermissionError>) -> Void) {