+++ CHANGELOG.md
@@ -1,17 +1,32 @@
 ## CHANGELOG
 
-### 3.x.x (2021-xx-xx)
+### 3.x.x (2021-05-DD)
+
+**SDK**
+
+- **Feature:** Added `rakuten.miniapp.user.action.SEND_MESSAGE` custom permission
+- **Feature:** Support Email addresses (`mailto:`) hyperlinks from a mini app. See [here](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a).
+- **Deprecated:** `getUniqueId()` in `MiniAppMessageDelegate` protocol is deprecated. You should use `getUniqueId(completionHandler:)` instead
+- **Deprecated:** `getContacts()` in `MiniAppUserInfoDelegate` protocol is deprecated. You should use `getContacts(completionHandler:)` instead
+- **Fixed:** Long messages were truncated in sample app
+
+**Sample App**
+- **Feature:** Custom contact name and email can be added now from the settings
+
+---
+### 3.2.0 (2021-04-22)
 
 **SDK**
 
 - **Feature:** Added chat messages capabilities
+- **Feature:** Standardized texts localization
 
 **Sample App**
 
 - **Feature:** Added chat messages screens with contact picker
 - **Fixed:** First launch screen metadata text was hidden
 
-
+---
 ### 3.1.0 (2021-04-02)
 
 **SDK**
@@ -61,7 +76,7 @@
 - **Feature** Added analytics for Mini App usage tracking
 - **Feature** Updated `getUserName` and `getProfilePhoto` interfaces to be asynchronous. Old methods are deprecated
 - **Feature** Updated `MiniApp().shared.create` interface to accept another optional query string parameter.
-- **Deprecated:** `requestPermission(permissionType:completionHandler:)` in `MiniAppMessageDelegate` protocol is deprecated. You should use `requestDevicePermission(permissionType:completionHandler:))` instead
+- **Deprecated:** `requestPermission(permissionType:completionHandler:)` in `MiniAppMessageDelegate` protocol is deprecated. You should use `requestDevicePermission(permissionType:completionHandler:)` instead
 - **Fixed** Geolocation was not updated after the first GPS fix
 
 **Sample App**
@@ -113,7 +128,7 @@
 - **Feature:** Added support for Orientation lock, that enables the mini app to lock `portrait` or `landscape` orientation for the mini-app. [Please check here](USERGUIDE.md#orientation-lock)
 - **Feature:** Added support to retrieve Access token and expiry date
 - **Feature:** Added default implementation in SDK for `requestCustomPermissions(permissions:miniAppTitle:completionHandler:)` [Please check here](USERGUIDE.md#request-custom-permission)
-- **Deprecated:** `requestCustomPermissions(permissions:completionHandler:)` in `MiniAppCallbackProtocol` protocol is deprecated. You should use `requestCustomPermissions(permissions:miniAppTitle:completionHandler:))` instead
+- **Deprecated:** `requestCustomPermissions(permissions:completionHandler:)` in `MiniAppCallbackProtocol` protocol is deprecated. You should use `requestCustomPermissions(permissions:miniAppTitle:completionHandler:)` instead
 
 **Sample App**